Move server ouside the main binary
