+++ .gitignore
@@ -4,7 +4,6 @@
 *.dll
 *.so
 *.dylib
-server
 
 # Test binary, build with `go test -c`
 *.test
