+++ cmd/url_shortener/url_shortener.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	server "github.com/tuttiq/go_url/server"
+)
+
+func main() {
+	server.StartServer()
+}
