+++ server/server.go
@@ -1,8 +1,8 @@
-package main
+package server
 
 import (
 	"fmt"
-	"io/ioutil"
+	// "io/ioutil"
 	"log"
 	"net/http"
 )
@@ -15,14 +15,16 @@ type Page struct {
 
 // function to load a page (any page)
 func loadPage(title string) (*Page, error) {
-	filename := title + ".html"
-
+	// filename := title + ".html"
 	// read file from view/filename.html
-	body, err := ioutil.ReadFile("views/" + filename)
+	// TODO: read from a real file like
+	// body, err := ioutil.ReadFile("views/" + filename)
 
-	if err != nil {
-		return nil, err
-	}
+	body := []byte("<html><body><h1>My Index Page</h1><div>Some body text</div></body></html>")
+
+	// if err != nil {
+	// 	return nil, err
+	// }
 
 	// return a Page object with the page title and html body
 	return &Page{Title: title, Body: body}, nil
@@ -34,7 +36,11 @@ func viewHandler(w http.ResponseWriter, r *http.Request) {
 	title := r.URL.Path[len("/view/"):]
 
 	// load page with the title requested
-	p, _ := loadPage(title)
+	p, err := loadPage(title)
+
+	if err != nil {
+		log.Println(err)
+	}
 
 	// parse the HTML body to string to write on the response
 	// TODO: not really using the page title for anything lol
@@ -46,7 +52,8 @@ func rootHandler(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintf(w, "Hi there, I love %s!", r.URL.Path[1:])
 }
 
-func main() {
+// StartServer starts the server and a minimal router
+func StartServer() {
 	http.HandleFunc("/", rootHandler)
 	http.HandleFunc("/view/", viewHandler)
 	log.Fatal(http.ListenAndServe(":8080", nil))
