+++ server/server_test.go
@@ -1,0 +1,7 @@
+package server
+
+import "testing"
+
+func TestStartServer(t *testing.T) {
+	// How to test a server being served?
+}
